LayoutTests/http/tests/media/media-source/video-media-source-state-changes-expected.txt

Tests MediaSource API state transitions.

webkitSourceState : closed
EVENT(webkitsourceopen) : open
onFirstSourceOpen
EVENT(webkitsourceended) : ended
onFirstSourceEnded
EVENT(playing)
triggerFirstSeek
EVENT(webkitsourceopen) : open
EVENT(seeking)
onFirstSeeking
EVENT(seeked)
onFirstSeeked
EVENT(seeking)
onSecondSeeking
EVENT(seeked)
onSecondSeeked
EVENT(webkitsourceclose) : closed
onFirstSourceClose
EVENT(emptied)
EVENT(webkitsourceopen) : open
onSecondSourceOpen
EVENT(webkitsourceended) : ended
onSecondSourceEnded
EVENT(playing)
triggerSecondSourceClose
EVENT(webkitsourceclose) : closed
onSecondSourceClose
EVENT(emptied)
EVENT(webkitsourceopen) : open
onThirdSourceOpen
END OF TEST
